requirements.txt
PyYAML==3.12
git+git://github.com/IanHarvey/bluepy
